requirements_win.txt
librosa==0.9.2
fairseq==0.12.2
Flask==2.1.2
Flask_Cors==3.0.10
gradio==3.4.1
numpy==1.20.0
playsound==1.3.0
PyAudio==0.2.12
pydub==0.25.1
pyworld==0.3.0
requests==2.28.1
scipy==1.7.3
sounddevice==0.4.5
SoundFile==0.10.3.post1
starlette==0.19.1
tqdm==4.63.0
scikit-maad
praat-parselmouth
onnx
onnxsim
onnxoptimizer
